feat: change file structure
