+++ packages/ts-newskit/src/slices/lead-story-1/styles.ts
@@ -6,11 +6,6 @@ import {
   Block
 } from 'newskit';
 
-export const ArticleDividerXL = styled(Divider)`
-  position: absolute;
-  height: 100%;
-  left: 50%;
-`;
 export const StyledDivider = styled(Divider)`
   ${getColorCssFromTheme('borderColor', 'interface050')};
 `;
